move unread indicator
